refactor: put universal logic to app
